+++ src/client/renderer/r_element.c
@@ -122,7 +122,7 @@ static int32_t R_SortParticles_Compare(const void *a, const void *b) {
 		const r_particle_t *bp = ((const r_particle_t *) be->element);
 
 		if (bp->type == ap->type)
-			return bp->image - ap->image;
+			return (int32_t) (intptr_t) (bp->image - ap->image);
 
 		return bp->type - ap->type;
 	}
